refactor: move test files to integration and spec
